cdk/pyproject.toml
[tool.poetry]
name = "serialconsole_config"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "serialconsole_config", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.143.0"
constructs = ">=10.0.0,<11.0.0"
cloudformation-cli = "^0.2.36"
cloudformation-cli-python-plugin = "^2.1.9"
setuptools = "^70.0.0"
gitpython = "^3.1.43"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
